fix(k8s/broker): move fake broker funcs into test files
