+++ k8s/broker/watch_broker_func_test.go
@@ -1,0 +1,16 @@
+package broker
+
+import (
+	"k8s.io/client-go/pkg/watch"
+)
+
+func newFakeWatchBrokerFunc(retErr error) (WatchBrokerFunc, *watch.FakeWatcher) {
+	fake := watch.NewFake()
+	fn := func(namespace string) (watch.Interface, error) {
+		if retErr != nil {
+			return nil, retErr
+		}
+		return fake, nil
+	}
+	return fn, fake
+}
